number format in Update progress.py
Updated line 275 to allow for the number format to be introduced in the "monitor" and "monitor_end" parameters.

In the current code, monitor = ("{count:,}/{total:,} [{percent:3.1%}] - processing "), throws an error. 

By adding these mods, the error is gone and the display show all the python formats available. e.g. below: 

|████████████████████████████████████████| 1,000/1,000 [100.00%] - ### COMPLETED ###  in 15.8s (63.41/s)

I hope it doesn't ruins something else in the code.